plsql_spring_boot/plsql_spring_boot/requirements.txt
python-dotenv
langchain-core
langchain-openai
chromadb
langchain-chroma
langchain-community
langchain-experimental